requirements.txt
langchain
langchain-community
python-dotenv
streamlit
unstructured
tiktoken
libmagic
python-magic
python-magic-bin
genai
chromadb
GoogleGenerativeAIEmbeddings
langchain-google-genai
langchain_google_genaipi